[python-frontend] move function_call_* files into function_call/ subdirectory
