+++ src/python-frontend/CMakeLists.txt
@@ -43,9 +43,9 @@ add_library(pythonfrontend STATIC
             cmath_lowering_policy.cpp
             complex_handler.cpp
             complex_handler_utils.cpp
-            function_call_expr.cpp
+            function_call/expr.cpp
             numpy_call_expr.cpp
-            function_call_builder.cpp
+            function_call/builder.cpp
             string_builder.cpp
             string_handler.cpp
             string_handler_utils.cpp
